feat(background-submission): :sparkles: Add proper background submission with queue and thread
Added a mechanism that takes the payload and puts that on queue and spawns 1 extra thread for background thread. That thread reads the data from queue, adds WAF threat score then appends to a list. Later that list is submitted as a batch to reduce network overhead.

http_logger, base_logger